⭐feat: 문법 수정 및 메인메뉴 라우터로 생성
