+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@popperjs/core": "2.11.5",
+    "@vue/cli-service": "^5.0.8",
     "bootstrap": "5.1.3",
     "pinia": "2.0.14",
     "prismjs": "1.28.0",
