+++ src/App.vue
@@ -1,20 +1,17 @@
 <template>
   <div id="app">
+    <MenuBar />
     <router-view></router-view>
   </div>
 </template>
 
 <script>
-// import MenuLinks from "./components/MenuLinks.vue";
-// import MainContent from "./components/MainContent.vue";
+import MenuBar from "@/components/MenuLinks.vue";
 
 export default {
   name: "App",
-  // components: {
-  //   MenuLinks,
-  //   MainContent
-  // }
-}
+  components: {MenuBar},
+};
 </script>
 
 <style>
