--- src/components/MainContent.vue
@@ -1,30 +1,0 @@
-<template>
-  <div class="main-content">
-    <div class="section section1">
-      <h1>작심백일</h1>
-      <p>이루고 싶은 목표가 무엇인가요?</p>
-    </div>
-    <div class="section section2">
-      <!-- lorem ipsum으로 대체 -->
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer nec odio. Praesent libero. Sed cursus ante
-        dapibus diam.</p>
-    </div>
-    <div class="section section3">
-      <!-- lorem ipsum으로 대체 -->
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer nec odio. Praesent libero. Sed cursus ante
-        dapibus diam.</p>
-    </div>
-  </div>
-</template>
-
-<script>
-export default {
-  data() {
-    return {
-
-    };
-  }
-};
-</script>
-
-<style src="../assets/css/Main.css" lang="css"></style>
