+++ src/main.js
@@ -1,5 +1,5 @@
 import { createApp } from "vue";
 import App from "./App.vue";
-import router from "./router";
+import router from "./router/router";
 
 createApp(App).use(router).mount("#app");
